Added page navigation functionality
Can now switch between MainPage and CanvasPage, no parameters being input yet

+++ WID/MainPage.xaml.cs
@@ -4,11 +4,16 @@ using System.IO;
 using System.Linq;
 using System.Numerics;
 using System.Threading.Tasks;
+using Windows.ApplicationModel;
+using Windows.ApplicationModel.Core;
 using Windows.Storage;
 using Windows.Storage.Pickers;
 using Windows.Storage.Streams;
 using Windows.UI.Input.Inking;
+using Windows.UI.ViewManagement;
+using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
+using Windows.UI.Xaml.Media.Animation;
 using WinRT;
 
 namespace WID
@@ -22,7 +27,14 @@ namespace WID
         public MainPage()
         {
             InitializeComponent();
+            SetTitlebar();
             LoadNotebooks();
+        }
+
+        private void SetTitlebar()
+        {
+            Window.Current.SetTitleBar(TitleBar);
+            tbAppTitle.Text = AppInfo.Current.DisplayInfo.DisplayName;
         }
 
         private async void LoadNotebooks()
@@ -55,9 +67,10 @@ namespace WID
             lvNotebooks.Items.Insert(index, new NotebookItem(newFolder.Name, true));
         }
 
-        private async void DeleteFolderOrFile(object sender, Windows.UI.Xaml.RoutedEventArgs e)
+        private void DeleteFolderOrFile(object sender, Windows.UI.Xaml.RoutedEventArgs e)
         {
             NotebookItem? nbItem = ((Button)sender).DataContext as NotebookItem;
+            if (nbItem == null) return;
             if (nbItem.IsFolder)
             {
                 Directory.Delete(notes.Path + "\\" + nbItem.Name.Replace("(Folder) ", ""), true);
@@ -67,6 +80,11 @@ namespace WID
                 File.Delete(notes.Path + "\\" + nbItem.Name);
                 lvNotebooks.Items.Remove(nbItem);
             }
+        }
+
+        private void OpenNote(object sender, Windows.UI.Xaml.RoutedEventArgs e)
+        {
+            Frame.Navigate(typeof(CanvasPage), null, new DrillInNavigationTransitionInfo());
         }
     }
 }
